requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.8.2
tenacity==8.5.0
pytest==8.3.3
psycopg2-binary==2.9.11
uvicorn==0.39.0
fastapi==0.128.8
APScheduler==3.11.2